Add Prolog CLI parser
